[IMP] point_of_sale: add support for Italy fiscal printer
Purpose
Since July 2019, it's mandatory in Italy to have your POS connected to a "fiscal printer", meaning a printer connected to the government. so the goal is to make POS compliant with Italian fiscal law integrating it specifically with the Epson Italian Fiscal Printers.

Updates Introduced:
- Support the Epson printer following operations:
    - print Fiscal Receipt
    - print Fiscal Invoice
    - print X Report
    - print Z Report
    - print XZ Report
    - re-print Receipt
    - refunds
    - discounts

- Adapt the POS tax calculations to be consistent with printer calculations

task-2092597

closes odoo/odoo#180458

Related: odoo/enterprise#70164
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>