[FIX] web: improve char field options documentation for Studio
Since dynamic_placeholder option was confusing in studio, compared to the recently
added placeholder_field option. And due to the poor usage and confusion it makes
besides mail templates or reports, we decided to remove documentation for those
options, and keep only the option that makes sense.

task-3901821

closes odoo/odoo#184657

Related: odoo/enterprise#72448
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>